replaced long with int

+++ source/Core/Calculation/SinglePhase/MultipleVoltageLevels/IPowerNetComputable.cs
@@ -4,7 +4,7 @@ namespace Calculation.SinglePhase.MultipleVoltageLevels
 {
     public interface IPowerNetComputable : IPowerNet
     {
-        IReadOnlyDictionary<long, NodeResult> CalculateNodeResults(out double relativePowerError);
+        IReadOnlyDictionary<int, NodeResult> CalculateNodeResults(out double relativePowerError);
         void CalculateAdmittanceMatrix(out AdmittanceMatrix matrix, out IReadOnlyList<string> nodeNames, out double powerScaling);
     }
 }
