+++ source/Core/Calculation/SinglePhase/MultipleVoltageLevels/PowerNetComputable.cs
@@ -14,7 +14,7 @@ namespace Calculation.SinglePhase.MultipleVoltageLevels
             _singleVoltagePowerNetFactory = singleVoltagePowerNetFactory;
         }
 
-        public IReadOnlyDictionary<long, NodeResult> CalculateNodeResults(out double relativePowerError)
+        public IReadOnlyDictionary<int, NodeResult> CalculateNodeResults(out double relativePowerError)
         {
             if (NodeGraph.FloatingNodesExist)
                 throw new InvalidDataException("there must not be a floating node");
@@ -31,7 +31,7 @@ namespace Calculation.SinglePhase.MultipleVoltageLevels
             if (nodeResults == null)
                 return null;
 
-            var nodeResultsWithId = new Dictionary<long, NodeResult>();
+            var nodeResultsWithId = new Dictionary<int, NodeResult>();
 
             foreach (var node in ExternalNodes)
             {
